feat: add SecurityFilter to authenticate users
The filter retries the JWT token taken from the request header. If the token is present, the filter then gets the subject of the token. The subject is the email address of the user. The filter then uses the email to find the user in the database. If the user is found the filter creates a UsernamePasswordAuthenticationToken and sets it in the SecurityContextHolder. This object is used by Spring Security to authenticate the user.