fix: modification the way a user reference is gotten
Before to find a user it was needed to supply an user id in the input data. Now a user is gotten by using the data in Authentication object. This way is safer since there's no way to insert invalid data.

The modifications occurred in the methods cadastrar and atualizarTopico